Move pipeline logic to dynacell/run.py and add YAML config support
Extract all orchestration code (run_all_fovs, SLURM submission, stage 1/2
logic) from debug_dynacell_preprocessing.py into dynacell/run.py. The
debug script is now a thin wrapper that runs tests then calls
run_from_config(). Add YAML config file with per-metric qc_thresholds,
SLURM parameters, and dataset settings.
